ARM: sunxi: quirk support
Currently, some hardware revisions of sunxi SoCs need special care on
some blocks because of hardware differences and/or bugs. Unfortunately,
it is unfeasible to account for these issues directly when writing the
device tree, as SoC revision can vary between different units of the
same device. This commit introduces a place to adjust DT compatibles
as needed to work around said issues before devices are probed. To
demonstrate usage, two quirks are added for the PLL2 and audio codec
on sun4i.

Signed-off-by: Emilio López <emilio@elopez.com.ar>